fix: move edit button inside comment
