+++ src/components/Feed/ActivityHeader/ActivityHeader.jsx
@@ -1,12 +1,7 @@
-import React, { useRef } from 'react'
 import * as Styled from './ActivityHeader.styled'
 import UserImage from '@components/UserImage'
 import ActivityReference from '../ActivityReference/ActivityReference'
 import ActivityDate from '../ActivityDate'
-import MenuContainer from '@components/Menu/MenuComponents/MenuContainer'
-import ActivityCommentMenu from '../ActivityCommentMenu/ActivityCommentMenu'
-import { toggleMenuOpen } from '@state/context'
-import { useDispatch } from 'react-redux'
 import { Icon } from '@ynput/ayon-react-components'
 
 const ActivityHeader = ({
@@ -15,16 +10,15 @@ const ActivityHeader = ({
   date,
   isRef,
   activity = {},
-  onDelete,
-  onEdit,
+
   children,
-  id,
+
   projectName,
   entityType,
   onReferenceClick,
   onReferenceTooltip,
 }) => {
-  const { referenceType, origin = {}, isOwner, activityType, versions = [], activityId } = activity
+  const { referenceType, origin = {}, activityType, versions = [], activityId } = activity
   const isMention = referenceType === 'mention'
 
   const isPublish = activityType === 'version.publish'
@@ -33,11 +27,6 @@ const ActivityHeader = ({
 
   const boldString = isMention ? `mentioned` : 'commented'
   const entityTypeString = isMention ? ` ${entityType} on` : 'on'
-
-  // open menu
-  const dispatch = useDispatch()
-  const handleToggleMenu = (menu) => dispatch(toggleMenuOpen(menu))
-  const moreRef = useRef()
 
   const noUser = activity.author?.deleted || !activity.author?.active
   return (
@@ -84,25 +73,6 @@ const ActivityHeader = ({
 
         {/* custom children, like status change */}
         {children}
-        {!isPublish && (
-          <Styled.Tools className={'tools'} ref={moreRef}>
-            {isOwner && onEdit && (
-              <Styled.ToolButton icon="edit_square" variant="text" onClick={onEdit} />
-            )}
-            {isOwner && (
-              <Styled.ToolButton
-                icon="more_horiz"
-                variant="text"
-                className="more"
-                onClick={() => handleToggleMenu(id)}
-              />
-            )}
-          </Styled.Tools>
-        )}
-
-        <MenuContainer id={id} target={moreRef.current}>
-          <ActivityCommentMenu onDelete={() => isOwner && onDelete()} />
-        </MenuContainer>
       </Styled.Body>
       <ActivityDate date={date} />
     </Styled.Header>
